requirements.txt
accelerate==1.9.0
diffusers==0.34.0
filelock==3.18.0
fsspec==2025.7.0
hf-xet==1.1.5
huggingface-hub==0.34.1
importlib-metadata==8.7.0
jinja2==3.1.6
markupsafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
pandas==2.3.3
pillow==11.3.0
psutil==7.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
regex==2024.11.6
safetensors==0.5.3
six==1.17.0
sympy==1.14.0
tokenizers==0.21.2
torch==2.7.1
torchvision==0.22.1
transformers==4.54.0
triton==3.3.1
typing-extensions==4.14.1
tzdata==2025.2
zipp==3.23.0